Lesson - v34 Finish update comment with Vue and add confirm alert for delete.

+++ resources/js/components/Comment.vue
@@ -3,30 +3,37 @@ export default {
 
     name: 'Comment',
 
-    props: ['comment-data'],
+    props: ['comment-data'],    // data about comment from blade
 
     data() {
         return {
-            editing: false
+            editing: false,
+            newText: ''
         }
     },
 
     mounted() {
-      console.log(this.commentData)
+      this.newText = this.commentData.text
     },
 
     methods: {
-        updateComment() {
-            axios.patch('/comments/' + this.commentData.id, {
-                text: 'lol lol hacked comment'
-            })
+        textChanged() {
+            this.newText = this.$refs.input.innerText
+        },
 
+        updateComment() {
             this.editing = false
+
+            axios.patch('/comments/' + this.commentData.id, {
+                text: this.newText
+            })
         },
 
         deleteComment() {
-            axios.delete('/comments/' + this.commentData.id)
-            this.$el.remove()
+            if (window.confirm('are you sure ???')) {
+                axios.delete('/comments/' + this.commentData.id)
+                this.$el.remove()
+            }
         }
     }
 }
